Move instance_event impl into event module
